requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.5.0
gunicorn==21.2.0
numpy==1.26.3
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
pymongo==4.6.1
pytz==2023.4
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
whitenoise==6.6.0
